refactor: useSearch hook 추상화
